reduced hyperparameter space of GIN and Baseline
training-loss-based early stopping nearly never is the best hyperparam config and even if it is, not by any significant margin. therefore it is removed, cutting evaluation times in half.

additional reason: the WL2 approach the results will be compared with, is also evaluated using only validation-based early stopping, so this change does not give an unfair advantage.